Move Layout related components to /components/SidebarLayout
