+++ src/components/SidebarLayout/Header/components/WalletIcon/icons/icon-authereum.png
