+++ src/components/SidebarLayout/Header/components/WalletIcon/icons/index.ts
