+++ src/components/SidebarLayout/Header/components/WalletIcon/index.tsx
