+++ src/components/SidebarLayout/SidebarLayout.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { Card } from '@gnosis.pm/safe-react-components'
 
-import Footer from '../Footer'
+import Footer from './Footer'
 
 const Grid = styled.div`
   height: 100%;
