+++ src/components/SidebarLayout/sidebarLayout.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from './Layout'
+import Layout from './SidebarLayout'
 import Sidebar from './Sidebar'
 
 export default {
