+++ src/components/layout/PageFrame/index.tsx
@@ -10,8 +10,8 @@ import InfoIcon from './assets/info.svg'
 import styles from './index.module.scss'
 
 import CookiesBanner from 'src/components/CookiesBanner'
-import Footer from 'src/components/Footer'
-import Header from 'src/components/Header'
+import Footer from 'src/components/SidebarLayout/Footer'
+import Header from 'src/components/SidebarLayout/Header'
 import Notifier from 'src/components/Notifier'
 import SidebarProvider from 'src/components/Sidebar'
 import Backdrop from 'src/components/layout/Backdrop'
